+++ mwcp/file_object.py
@@ -244,14 +244,13 @@ class FileObject(object):
             )
             self._outputted_file = True
 
-    def run_kordesii_decoder(self, decoder_name: str, warn_no_strings=True, decoderdir=None):
+    def run_kordesii_decoder(self, decoder_name: str, warn_no_strings=True):
         """
         Run the specified kordesii decoder against the file data.  The reporter object is returned
         and can be accessed as necessary to obtain output files, etc.
 
         :param decoder_name: name of the decoder to run
         :param warn_no_strings: Whether to produce a warning if no string were found.
-        :param decoderdir: Custom decoder directory to use instead of the default.
 
         :return: Instance of the kordesii_reporter.
 
@@ -265,7 +264,10 @@ class FileObject(object):
             return self._kordesii_cache[decoder_name]
 
         logger.info(f"Running {decoder_name} kordesii decoder on file {self.file_name}.")
-        kordesii_reporter = kordesii.Reporter(decoderdir=decoderdir, base64outputfiles=True)
+        # Ensure decoderdir sources are populated
+        kordesii.register_entry_points()
+
+        kordesii_reporter = kordesii.Reporter(base64outputfiles=True)
 
         kordesii_reporter.run_decoder(decoder_name, data=self.file_data, log=True)